drilldown: update query result for multiselect
Rather than filtering down each column so that it's equal to the query
parameters, calculate the set that would be returned if that column
weren't a constraint. This means that you can select a commit, and a
metric, and see a list of other commits you could compare to along
that metric, for example.

TODO: actually support multiple selections in a query

Change-Id: I1d8afc2fdf58e170909a50f13af5218f80702a30