treehandler: modified for file/commit hierarchies
Switched to identifying nodes by id (rather than the name a user
sees), and specified behavior when a parent node is selected. If a
parent node is selected, then all children are checked. If a parent
node is unchecked, then all children are unchecked. Note that this
javascript file expects all parent nodes (nodes it children) to have
an id that starts with a underscore character when recieved from
drilldown.

Change-Id: I2ecba97fbef83c81dd58a1d8342068fafed54718