Added formatting in drilldown for the UI
Added a formatting function to drilldown. Now results are passed to
the UI with hierarchies expressed (and display names are used for
viewing). The formatting function queries the memcache for additional
data (such as display name). To avoid circular importing, the db
model definitions were placed into a separate file.

Change-Id: Ifbafda237c381172feb71091eb193162bafb795e