commit-viewer: query metrics from datastore
Instead of reconstructing the contents of the datastore from drilldown,
get the list of available metrics by querying the CodecMetricIndex.
This provides a big speed boost, both by reducing the time to get the
list of metrics, and reducing the number of operations required to
get the data for those metrics, by better batching.

Change-Id: I9eac974ed1b9cd96fd6faa4b1d49f340f2e56335